python: move BIP32/BIP39 key-derivation helpers into centurymetadata.bip39
bip39_to_seed/_bip32_master/_bip32_child_h/_H/derive_cm_keys lived only
in examples/centurytool.py, undocumented as package-private. Move them
into centurymetadata.bip39 (alongside the checksum validator) so the
server's forthcoming known-keys module can reuse the same derivation
code instead of duplicating it. Pure refactor, no behavior change.
