+++ examples/centurytool.py
@@ -1,63 +1,11 @@
 #! /usr/bin/python3
 import centurymetadata
+from centurymetadata.bip39 import bip39_to_seed, derive_cm_keys
 import secp256k1
 import argparse
-import hashlib
-import hmac
 import json
 import requests
 import sys
-
-
-# ── BIP39 / BIP32 helpers ────────────────────────────────────────────────────
-
-SECP256K1_ORDER = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141
-
-
-def bip39_to_seed(mnemonic: str, passphrase: str = "") -> bytes:
-    """BIP39: derive 64-byte seed from mnemonic (no wordlist validation)."""
-    return hashlib.pbkdf2_hmac(
-        'sha512',
-        mnemonic.strip().encode('utf-8'),
-        ('mnemonic' + passphrase).encode('utf-8'),
-        2048,
-    )
-
-
-def _bip32_master(seed: bytes) -> tuple:
-    """BIP32 master key. Returns (privkey_bytes, chain_code_bytes)."""
-    I = hmac.new(b'Bitcoin seed', seed, hashlib.sha512).digest()
-    return I[:32], I[32:]
-
-
-def _bip32_child_h(privkey: bytes, chain: bytes, index: int) -> tuple:
-    """BIP32 hardened child (index must already have 0x80000000 set)."""
-    data = b'\x00' + privkey + index.to_bytes(4, 'big')
-    I = hmac.new(chain, data, hashlib.sha512).digest()
-    ki = (int.from_bytes(I[:32], 'big') + int.from_bytes(privkey, 'big')) % SECP256K1_ORDER
-    return ki.to_bytes(32, 'big'), I[32:]
-
-
-def _H(i: int) -> int:
-    return i | 0x80000000
-
-
-def derive_cm_keys(seed: bytes, n: int = 0) -> tuple:
-    """Derive centurymetadata keys for slot N.
-    Path: m / 0x44315441' / N' / {0,1,2,3}'
-      0' = writer secp256k1   1' = reader secp256k1
-      2' = writer ML-KEM seed 3' = reader ML-KEM seed
-    Returns (writer_secp_bytes, reader_secp_bytes, writer_mlkem_seed, reader_mlkem_seed).
-    """
-    PURPOSE = 0x44315441
-    k, c = _bip32_master(seed)
-    k, c = _bip32_child_h(k, c, _H(PURPOSE))
-    k, c = _bip32_child_h(k, c, _H(n))
-    w_secp, _ = _bip32_child_h(k, c, _H(0))
-    r_secp, _ = _bip32_child_h(k, c, _H(1))
-    w_mlkem, _ = _bip32_child_h(k, c, _H(2))
-    r_mlkem, _ = _bip32_child_h(k, c, _H(3))
-    return w_secp, r_secp, w_mlkem, r_mlkem
 
 
 # ── Fetch helper ─────────────────────────────────────────────────────────────
